+++ src/swganh/connection/connection_service.cc
@@ -276,7 +276,7 @@ void ConnectionService::HandleClientIdMsg_(std::shared_ptr<ConnectionClient> cli
 
     ClientPermissionsMessage client_permissions;
     client_permissions.galaxy_available = kernel()->GetServiceDirectory()->galaxy().status();
-    client_permissions.available_character_slots = character_service()->GetMaxCharacters(account_id);
+    client_permissions.available_character_slots = static_cast<uint8_t>(character_service()->GetMaxCharacters(account_id));
     // @TODO: Replace with configurable value
     client_permissions.unlimited_characters = 0;
